Modifications for using user-pathnames
When moving the downloaded files from the temporary directory to the final location, a pattern-check was performed to see if a specific time-structure (/2017/05/03/) was required. This pattern used 'yy' for year, 'mm' for month, 'dd' for day. However when storing data into my personal directory /home/jtimmer/, this approach wrongly assumed that a month-value needed to be used... To correct for this, we now use MM for month, YY for year, DD for day. 

This also requires the data_stores.yml (and it's template (vm_data_stores.yml) in the vm_support module to be updated).